// frontend/src/shared/lib/api/model/startUmapClusterResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { StartUmapClusterResponseResult } from './startUmapClusterResponseResult';
import type { StartUmapClusterResponseJob } from './startUmapClusterResponseJob';

export interface StartUmapClusterResponse {
  result: StartUmapClusterResponseResult;
  job: StartUmapClusterResponseJob;
}